requirements.txt
asgiref==3.5.2
certifi==2022.6.15
Django==4.0.5
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
pycurl==7.45.1